requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.34.0
Django==4.2.5
djangorestframework==3.14.0
mysqlclient==2.2.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
<<<<<<< HEAD
<<<<<<< HEAD
urllib3==1.26.16
=======
urllib3==1.26.16
>>>>>>> 5dec6a69ca57e6779e69eb9872c9c3600e6fc30f
=======
urllib3==1.26.16
>>>>>>> 61f328dee497fbcbfd352ac01075f5d666a17e4b
